discard interned OsStr/Path impls
Implementation of OsStr is tightly coupled with underlying OS.
So they're not intended to make conversion from/to &[u8], especially for
windows, which uses possibly-ill-formed UTF-16 as string format, though
OsString for it internally uses WTF-8 format which is byte-aligned.

Path is just a thin wrapper of OsStr so also discarded.

It's possible to make them internable with separated pool for OsStr. But
I don't think it's worth to implement now.